refactor: Tab 상태 전용 컨텍스트를 역할에 맞게 이름 변경
